+++ test/document.test.js
@@ -3791,6 +3791,27 @@ describe('document', function() {
       done();
     });
 
+    it('setting to discriminator (gh-4935)', function(done) {
+      var Buyer = db.model('gh4935_0', new Schema({
+        name: String,
+        vehicle: { type: Schema.Types.ObjectId, ref: 'gh4935' }
+      }));
+      var Vehicle = db.model('gh4935', new Schema({ name: String }));
+      var Car = Vehicle.discriminator('gh4935_1', new Schema({
+        model: String
+      }));
+
+      var eleanor = new Car({ name: 'Eleanor', model: 'Shelby Mustang GT' });
+      var nick = new Buyer({ name: 'Nicolas', vehicle: eleanor });
+
+      assert.ok(!!nick.vehicle);
+      assert.ok(nick.vehicle === eleanor);
+      assert.ok(nick.vehicle instanceof Car);
+      assert.equal(nick.vehicle.name, 'Eleanor');
+
+      done();
+    });
+
     it('modify multiple subdoc paths (gh-4405)', function(done) {
       var ChildObjectSchema = new Schema({
         childProperty1: String,